+++ sdk/resourcemanager/mysql/armmysqlflexibleservers/constants.go
@@ -11,8 +11,35 @@ package armmysqlflexibleservers
 
 const (
 	moduleName    = "armmysqlflexibleservers"
-	moduleVersion = "v1.1.0"
+	moduleVersion = "v1.2.0-beta.1"
 )
+
+type AdministratorName string
+
+const (
+	AdministratorNameActiveDirectory AdministratorName = "ActiveDirectory"
+)
+
+// PossibleAdministratorNameValues returns the possible values for the AdministratorName const type.
+func PossibleAdministratorNameValues() []AdministratorName {
+	return []AdministratorName{
+		AdministratorNameActiveDirectory,
+	}
+}
+
+// AdministratorType - Type of the sever administrator.
+type AdministratorType string
+
+const (
+	AdministratorTypeActiveDirectory AdministratorType = "ActiveDirectory"
+)
+
+// PossibleAdministratorTypeValues returns the possible values for the AdministratorType const type.
+func PossibleAdministratorTypeValues() []AdministratorType {
+	return []AdministratorType{
+		AdministratorTypeActiveDirectory,
+	}
+}
 
 // ConfigurationSource - Source of the configuration.
 type ConfigurationSource string
@@ -205,6 +232,22 @@ func PossibleReplicationRoleValues() []ReplicationRole {
 		ReplicationRoleNone,
 		ReplicationRoleReplica,
 		ReplicationRoleSource,
+	}
+}
+
+// ResetAllToDefault - Whether to reset all server parameters to default.
+type ResetAllToDefault string
+
+const (
+	ResetAllToDefaultFalse ResetAllToDefault = "False"
+	ResetAllToDefaultTrue  ResetAllToDefault = "True"
+)
+
+// PossibleResetAllToDefaultValues returns the possible values for the ResetAllToDefault const type.
+func PossibleResetAllToDefaultValues() []ResetAllToDefault {
+	return []ResetAllToDefault{
+		ResetAllToDefaultFalse,
+		ResetAllToDefaultTrue,
 	}
 }